refactor: add controller dir
- rename HelloController.java
- add controller dir
