+++ src/main/java/com/galewings/controller/GaleWingsController.java
@@ -1,4 +1,4 @@
-package com.galewings;
+package com.galewings.controller;
 
 import com.galewings.entity.SiteFeedCount;
 import com.galewings.repository.SiteRepository;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
-public class HelloController {
+public class GaleWingsController {
 
   @Autowired
   SqlManager sqlManager;
