+++ src/main/java/com/galewings/controller/SettingsController.java
@@ -1,4 +1,4 @@
-package com.galewings;
+package com.galewings.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.galewings.dto.input.SettingsUpdateForm;
